refactored bootcamp controller so advanced filter is generic and could be reused for other models as well
