+++ routes/bootcamps.js
@@ -9,6 +9,10 @@ const {
   bootcampPhotoUpload
 } = require('../controllers/bootcamps');
 
+const Bootcamp = require('../models/Bootcamp');
+
+const advancedResults = require('../middleware/advancedResults');
+
 // Include other resource routers
 const courseRouter = require('./courses');
 
@@ -21,7 +25,10 @@ router.route('/radius/:zipcode/:distance').get(getBootcampsInRadius);
 
 router.route('/:id/photo').put(bootcampPhotoUpload);
 
-router.route('/').get(getBootcamps).post(createBootcamp);
+router
+  .route('/')
+  .get(advancedResults(Bootcamp, 'courses') ,getBootcamps)
+  .post(createBootcamp);
 
 router
   .route('/:id')
